// src/api/model/getFriendsNetworkCircleSize.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * DunbarHorizon API 명세서
 * DunbarHorizon 백엔드 API
 * OpenAPI spec version: 1.0.0
 */

export type GetFriendsNetworkCircleSize = typeof GetFriendsNetworkCircleSize[keyof typeof GetFriendsNetworkCircleSize];


export const GetFriendsNetworkCircleSize = {
  SUPPORT: 'SUPPORT',
  SYMPATHY: 'SYMPATHY',
  KINSHIP: 'KINSHIP',
  DUNBAR: 'DUNBAR',
} as const;